Adds contoller file for route functions
